1.2.Adding a Comment Entity{Creating the Comment Entity: php bin/console make:entity}
Creating the Comment Entity
In the src/Entity directory, the only entity we have so far is Article
So before we can talk about relationships, we first need to build a Comment entity. We could create this by hand, but the generator is so much nicer:

Open a new terminal tab and run:

php bin/console make:entity
Name the entity Comment. Then, for the fields, we need one for the author and one for the actual comment. Add authorName as a string field. And yea, someday, we might have a User table. And then, this could be a relationship to that table. But for now, keep it as a simple string.

Next, add content as a text field, and also say no to nullable. Hit enter one more time to finish up.