1.4.Adding a Comment Entity{Creating the Comment Entity: php bin/console make:migration && php bin/console doctrine:migrations:migrate}
Now head back to your terminal and run:

php bin/console make:migration
When that finishes, go find the new file. We just want to make sure that this doesn't contain any surprises. For example, if you're working on multiple branches, then your database may be out-of-sync before you run make:migration. If that happens, the migration file would contain extra changes that you'll want to remove. In this case, it looks great.Go back to your terminal and, migrate!

php bin/console doctrine:migrations:migrate
Perfect! We have an article table and now a comment table. But, they are not friends yet. Time to add a relation!