2.2.Adding the ManyToOne Relation{Generating the Migration: php bin/console make:migration && php bin/console doctrine:migrations:migrate}
Generating the Migration
Enough talking! Let's finally generate the migration. Find your terminal and run:

php bin/console make:migration
Go back to your editor and open that new migration file. Woh! Awesome! The end-result is super simple: it adds a new article_id column to the comment table along with a foreign key constraint to the article's id column.
So even though, in Comment, we called the property article In the database, this creates an article_id column! Ultimately, the database looks exactly like we expected in the beginning! But in PHP, guess what? When we set this article property, we will set an entire Article object on it - not the Article's ID. More about that next.

The migration looks prefect. So find your terminal, and run it!

php bin/console doctrine:migrations:migrate
Ok, time to create a Comment object and learn how to relate it to an Article.