3.1.Saving Relations{JoinColumn & Required Foreign Key Columns}
Our Comment entity has an article property and an article_id column in the database.
So, the question now is: how do we actually populate that column? How can we relate a Comment to an Article?
The answer is both very easy, and also, quite possibly, at first, weird! Open up the ArticleFixtures class. Let's hack in a new comment object near the bottom: $comment1 = new Comment()
Then, $comment1->setAuthorName(), and we'll go copy our favorite, always-excited astronaut commenter: Mike Ferengi. Then, $comment1->setContent(), and use one of our hardcoded comments.
Perfect! Because we're creating this manually, we need to persist it to Doctrine. At the top, use the $manager variable. Then, $manager->persist($comment1).
If we stop here, this is a valid Comment... but it is NOT related to any article. In fact, go to your terminal, and try the fixtures:

php bin/console doctrine:fixtures:load

JoinColumn & Required Foreign Key Columns
Boom! It fails with an integrity constraint violation:

  Column article_id cannot be null

It is trying to create the Comment, but, because we have not set the relation, it doesn't have a value for article_id.
Oh, and also, in Comment, see this JoinColumn with nullable=false? That's the same as having nullable=false on a property: it makes the article_id column required in the database. Oh, but, for whatever reason, a column defaults to nullable=false, and JoinColumn defaults to the opposite: nullable=true.