4.1.Fetching Relations
Yes! Each Article is now related to two comments in the database. So, on the article show page, it's time to get rid of this hardcoded stuff and, finally, query for the true comments for this Article.
In src/Controller, open ArticleController and find the show() action
This renders a single article. So, how can we find all of the comments related to this article? Well, we already know one way to do this.
Remember: whenever you need to run a query, step one is to get that entity's repository. And, surprise! When we generated the Comment class, the make:entity command also gave us a new CommentRepository
Thanks MakerBundle!
Get the repository by adding a CommentRepository argument. Then, let's see, could we use one of the built-in methods? Try $comments = $commentRepository->findBy(), and pass this article set to the entire $article object. Dump these comments and die.
Then, find your browser and, try it! Yes! It returns the two Comment objects related to this Article!
So, the weird thing is that, once again, you need to stop thinking about the columns in your tables, like article_id, and only think about the properties on your entity classes. That's why we use 'article' => . Of course, behind the scenes, Doctrine will make a query where article_id = the id from this Article. But, in PHP, we think all about objects.