4.2.Fetching Relations{Fetching Comments Directly from Article}
Fetching Comments Directly from Article
But there is a much simpler way! When we generated the relationship, it asked us if we wanted to add an optional comments property to the Article class, for convenience. We said yes! And thanks to that, we can literally say $comments = $article->getComments(). Dump $comments again.
Oh, and now, we don't need the CommentRepository anymore.
Cool.