4.3.Fetching Relations{Lazy Loading}
Lazy Loading
Head back to your browser and, refresh! It's the exact same as before. Wait, what? What's this weird PersistentCollection thing?
Here's what's going on. When Symfony queries for the Article, it only fetches the Article data: it does not automatically fetch the related Comments. And, for performance, that's great! We may not even need the comment data! But, as soon as we call getComments() and start using that, Doctrine makes a query in the background to go get the comment data.
This is called "lazy loading": related data is not queried for until, and unless, we use it. To make this magic possible, Doctrine uses this PersistentCollection object. This is not something you need to think or worry about: this object looks and acts like an array.

To prove it, let's foreach over $comments as $comment and dump each $comment inside. Put a die at the end
Try it again! Boom! Two Comment objects!