4.4.Fetching Relations{Fetching the Comments in the Template}
Fetching the Comments in the Template
Back in the controller, we no longer need these hard-coded comments. In fact, we don't even need to pass comments into the template at all! That's because we can call the getComments() method directly from Twig!
Remove all of the comment logic. And then, jump into templates/article/show.html.twig. Scroll down a little... ah, yes! First, update the count: article.comments|length.
Easy! Then, below, change the loop to use for comment in article.comments. And because each comment has a dynamic author, print that with {{ comment.authorName }}. And the content is now comment.content.
Oh, and, because each comment has a createdAt, let's print that too, with our trusty ago filter.
Love it! Let's try it! Go back, refresh and... yes! Two comments, from about 17 minutes ago. And, check this out: on the web debug toolbar, you can see that there are two database queries. The first query selects the article data only. And the second selects all of the comment data where article_id matches this article's id - 112. This second query doesn't actually happen until we reference the comments from inside of Twig.
That laziness is a key feature of Doctrine relations.
Next, it's time to talk about the subtle, but super-important distinction between the owning and inverse sides of a relation.