6.2.Fixture References & Relating Objects{php bin/console make:fixture}
Flip back to your editor and open CommentFixture file.
In the last tutorial, we made a cool base class with some extra shortcuts. Extend BaseFixture. Then, instead of load, we now need loadData(), and it should be protected. Remove the extra use statement on top.
Thanks to our custom base class, we can create a bunch of comments easily with $this->createMany(), passing it Comment::class, 100, and then a callback that will receive each 100 Comment objects.
Inside, let's use Faker - which we also setup in the last tutorial - to give us awesome, fake data. Start with $comment->setContent(). I'll use multiple lines.
Now, if $this->faker->boolean, which will be a random true or false, then either generate a random paragraph: $this->faker->paragraph, or generate two random sentences. Pass true to get this as text, not an array.
Cool! Next, for the author, we can use $comment->setAuthor() with $this->faker->name, to get a random person's name.
By the way, all of these faker functions are covered really well in their docs. I'm seriously not just making them up.
Finally, add $comment->setCreatedAt() with $this->faker->dateTimeBetween() from -1 months to -1 seconds. That'll give us much more interesting data.