8.1.OrderBy & fetch EXTRA_LAZY
It's great that each article now has real, dynamic comments at the bottom. But... something isn't right: the comments are in a, kind of, random order. Actually, they're just printing out in whatever order they were added to the database. But, that's silly! We need to print the newest comments on top, the oldest at the bottom.

How can we do this? Check out the template. Hmm, all we're doing is calling article.comments:

 87 lines  templates/article/show.html.twig
<div class="container">
    <div class="row">
        <div class="col-sm-12">
            <div class="show-article-container p-3 mt-4">
                <div class="row">
                    <div class="col-sm-12">
                        {% for comment in article.comments %}
                        <div class="row">
                            <div class="col-sm-12">
                                <img class="comment-img rounded-circle" src="{{ asset('images/alien-profile.png') }}">
                                <div class="comment-container d-inline-block pl-3 align-top">
                                    <span class="commenter-name">{{ comment.authorName }}</span>
                                    <small>about {{ comment.createdAt|ago }}</small>
                                    <br>
                                    <span class="comment"> {{ comment.content }}</span>
                                    <p><a href="#">Reply</a></p>
                                </div>
                            </div>
                        </div>
                        {% endfor %}
                    </div>
                </div>
            </div>
        </div>
    </div>
</div>
Which is the getComments() method on Article:

 202 lines  src/Entity/Article.php
class Article
{
    /**
     * @return Collection|Comment[]
     */
    public function getComments(): Collection
    {
        return $this->comments;
    }
}
The great thing about these relationship shortcut methods is that.... they're easy! The downside is that you don't have a lot of control over what's returned, like, the order of this article's comments.

Well... that's not entirely true. We can control the order, and I'll show you how. Actually, we can control a lot of things - but more on that later.