8.4.OrderBy & fetch EXTRA_LAZY{Fetch EXTRA_LAZY}
As soon as you call getComments() and use that data, Doctrine, by default, makes a query at that moment to get all of the comment data, even if you eventually only need to count that data.
But, we can control this. In Article, at the end of the OneToMany annotation, add fetch="EXTRA_LAZY"
Now, go back to the page and refresh. We still have six queries, but go look at them. Awesome! Instead of selecting all of the comment data, they are super-fast COUNT queries!
Here's how this works: if you set fetch="EXTRA_LAZY", and you simply count the result of $article->getComments(), then instead of querying for all of the comments, Doctrine does a quick COUNT query.
Awesome, right! You might think that it's so awesome that this should always be the way it works! But, there is one situation where this is not ideal. And actually, we have it! Go to the article show page.

Here, we count the comments first, and then we loop over them:

 87 lines  templates/article/show.html.twig
{% block body %}
<div class="container">
    <div class="row">
        <div class="col-sm-12">
            <div class="show-article-container p-3 mt-4">
                <div class="row">
                    <div class="col-sm-12">
                        <h3><i class="pr-3 fa fa-comment"></i>{{ article.comments|length }} Comments</h3>
                        {% for comment in article.comments %}
                        {% endfor %}
                    </div>
                </div>
            </div>
        </div>
    </div>
</div>
{% endblock %}
Look at the profiler now. Thanks to EXTRA_LAZY, we have an extra query! It counts the comments... but then, right after, it queries for all of them anyways. Before we were using EXTRA_LAZY, this count query didn't exist.
So, sorry people, like life, everything is a trade-off. But, it's still probably a net-win for us. But as always, don't prematurely optimize. Deploy first, identify performance issues, and then solve them.