9.1.Giving the Comments an isDeleted Flag
I want to show you a really cool, really powerful feature. But, to do that, we need to give our app a bit more depth. We need to make it possible to mark comments as deleted. Because, honestly, not all comments on the Internet are as insightful and amazing as the ones that you all add to KnpUniversity. You all are seriously the best! But, instead of actually deleting them, we want to keep a record of deleted comments, just in case.

Adding Comment.isDeleted Field
Here's the setup: go to your terminal and run:

php bin/console make:entity
We're going to add a new field to the Comment entity called isDeleted. This will be a boolean type and set it to not nullable in the database.