9.2.Giving the Comments an isDeleted Flag{php bin/console make:migration && php bin/console doctrine:migrations:migrate}
Make the migration:

php bin/console make:migration
And, you know the drill: open that migration to make sure it doesn't contain any surprises.
Oh, this is cool: when you use a boolean type in Doctrine, the value on your entity will be true or false, but in the database, it stores as a tiny int with a zero or one.

This looks good, so move back and.... migrate!

php bin/console doctrine:migrations:migrate