9.3.Giving the Comments an isDeleted Flag{Updating the Fixtures}
Updating the Fixtures
We're not going to create an admin interface to delete comments, at least, not yet. Instead, let's update our fixtures so that it loads some "deleted" comments. But first, inside Comment, find the new field and... default isDeleted to false.
Any new comments will not be deleted.
Next, in CommentFixture, let's say $comment->setIsDeleted() with $this->faker->boolean(20).
So, out of the 100 comments, approximately 20 of them will be marked as deleted.
Then, to make this a little bit obvious on the front-end, for now, open show.html.twig and, right after the date, add an if statement: if comment.isDeleted, then, add a close, "X", icon and say "deleted"
Find your terminal and freshen up your fixtures:

php bin/console doctrine:fixtures:load
When that finishes, move back, refresh... then scroll down. Let's see... yea! Here's one: this article has one deleted comment.
We printed this "deleted" note mostly for our own benefit while developing. Because, what we really want to do is, of course, not show the deleted comments at all!
But... hmm. The problem is that, to get the comments, we're calling article.comments. Which means we're calling Article::getComments(). This is our super-handy, super-lazy shortcut method that returns all of the comments. Dang!