9.4.Giving the Comments an isDeleted Flag{Hiding Deleted Comments}
Hiding Deleted Comments
Now we need a way to return only the non-deleted comments. Is that possible?
Yes! One option is super simple. Instead of using article.comments, we could go into ArticleController, find the show() action, create a custom query for the Comment objects we need, pass those into the template, then use that new variable. When the shortcut methods don't work, always remember that you don't need to use them.
But, there is another option, it's a bit lazier, and a bit more fun.

Creating Article::getNonDeletedComments()
Open Article and find the getComments() method. Copy it, paste, and rename to getNonDeletedComments(). But, for now, just return all of the comments. Then, in the show template, use this new field: in the loop, article.nonDeletedComments. And, further up, when we count them, also use article.nonDeletedComments.
Let's refresh to make sure this works so far. No errors, but, of course, we are still showing all of the comments.