9.5.Giving the Comments an isDeleted Flag{Filtering Deleted Comments in Article::getNonDeletedComments()}
Filtering Deleted Comments in Article::getNonDeletedComments()
Back in Article, how can we change this method to filter out the deleted comments? Well, there is a lazy way, which is sometimes good enough. And an awesome way! The lazy way would be to, for example, create a new $comments array, loop over $this->getComments(), check if the comment is deleted, and add it to the array if it is not. Then, at the bottom, return a new ArrayCollection of those comments:
$comments = [];

foreach ($this->getComments() as $comment) {
    if (! $comment->getIsDeleted()) {
        $comments[] = $comment;
    }
}

return new ArrayCollection($comments);
Simple! But... this solution has a drawback... performance! Let's talk about that next, and, the awesome fix.