11.2.Twig Block Tricks{Building the Comment Admin Template}
Building the Comment Admin Template
Let's open that template and get to work!
This already overrides the title block, which is cool! Change it to say "Manage Comments". Then, delete all the body code.
To make the page look nice, open show.html.twig: we need to steal some markup from this. Copy the first 6 divs. Back in index.html.twig, paste, close each of those 6 divs and... back in the middle, add Manage Comments.
Try that again in your browser: refresh. Ok! Those 6 divs give us this white box that you also see on the article show page.
Hmm. If you think about it, it's probably going to be really common for us to want a page where we have some nice margin and a white box. The homepage doesn't need this, but, I bet a lot of internal pages will.
So, needing to duplicate these six divs on each page is pretty lame. Fortunately, Twig comes to the rescue! Go Twig! We can isolate this markup into a new base layout.